Add many more syscall, fix many more bugs.

+++ include/device.h
@@ -23,8 +23,8 @@ int dev_register(struct device *dev, struct file_calls *calls, int type);
 	        ((unsigned)( (((x)>>12) & 0xffffff00) | ((x) & 0x000000ff) ))
 
 #define makedev(x,y) ( \
-		        (((x)&0xfffff000ULL) << 32) | \
-		        (((x)&0x00000fffULL) << 8) | \
+		        ((((unsigned long)x)&0xfffff000ULL) << 32) | \
+		        ((((unsigned long)x)&0x00000fffULL) << 8) | \
 		        (((y)&0xffffff00ULL) << 12) | \
 		        (((y)&0x000000ffULL)) )
